Create RasterRDD Mask Package in spark Project
