+++ spark/src/main/scala/geotrellis/spark/package.scala
@@ -35,6 +35,7 @@ import scala.reflect.ClassTag
 
 package object spark
     extends buffer.Implicits
+    with mask.Implicits
     with merge.Implicits
     with reproject.Implicits
     with tiling.Implicits
@@ -111,6 +112,9 @@ package object spark
 
   implicit class withRasterRDDMethods[K](val self: RasterRDD[K])(implicit val keyClassTag: ClassTag[K])
     extends RasterRDDMethods[K]
+
+  implicit class withRasterRDDMaskMethods[K: SpatialComponent: ClassTag](val self: RasterRDD[K])
+      extends mask.RasterRDDMaskMethods[K]
 
   implicit class withMultiBandRasterRDDMethods[K](val self: MultiBandRasterRDD[K])(implicit val keyClassTag: ClassTag[K])
     extends MultiBandRasterRDDMethods[K]
